refactor: move the types folder at root level
